+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.3.1
+
+* Default formatter: add support for printing `cause` of [JsonUnsupportedObjectError]
+
+## 1.3.0
+
+* Support for exception chaining.
+* Allow closing of subscriptions without disposing. `detachFromLoggers`
+
 ## 1.2.0+1
 
 * Graylog: Send timestamp with decimal places.